% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra_man.R
\name{dijkstra_man}
\alias{dijkstra_man}
\title{Title}
\usage{
dijkstra_man(graph, init_node)
}
\arguments{
\item{graph}{the input dataframe which represents a network}

\item{init_node}{the node for which the distance to other nodes must be calculated}
}
\value{
returns a vector with distances to other nodes
}
\description{
Title
}
\examples{
wiki_graph <-
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra_man(wiki_graph, 1)
}
