% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Title euclidean() returns the gcd}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{first parameter}

\item{b}{second parameter}
}
\value{
the gcd
}
\description{
Title euclidean() returns the gcd
}
\examples{
euclidean(100,1000)
}
