% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A graph dataset is present here}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{source node}
  \item{v2}{destination node}
  \item{w}{weight or distance}
  ...
}}
\usage{
wiki_graph
}
\description{
A dataset containing the source, destination and weight of each node
}
\keyword{datasets}
